services/intel-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
requests==2.32.3
apscheduler==3.10.4
pyyaml==6.0.2
